Struct float::Float

pub struct Float {
    // some fields omitted
}

Methods

impl Float

fn sqrt(self) -> Float

impl Float

fn zero(p: u32) -> Float

fn neg_zero(p: u32) -> Float

fn max(p: u32) -> Float

fn min(p: u32) -> Float

fn min_positive(p: u32) -> Float

fn infinity(p: u32) -> Float

fn neg_infinity(p: u32) -> Float

fn nan(p: u32) -> Float

fn next_above(self) -> Float

fn next_below(self) -> Float

fn next_toward(self, target: &Float) -> Float

fn rand<R: Rng>(r: &mut R, p: u32) -> Float

Generate a float in [0, 1), with the same distribution as generating an integer in [0, 2**p) and dividing by 2**p. (i.e. pretty close to uniform, but the smallest bits aren't quite right).

fn abs(self) -> Float

fn sign(&self) -> Option<Sign>

fn negate(&mut self)

fn precision(&self) -> u32

fn with_precision(self, prec: u32) -> Float

fn mul_exp2(self, exp: i64) -> Float